+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 - Automate release procedure.
+- Add contributing guidelines.
 
 ## 1.5.4 - 2019-07-23